[linux] move fbdup code into linux.c #2108
